Fixed issue with linking to DPP
Usually we would call dpp::dpp as a link target, but due to an issue
with DPP it is trying to enable coroutine support even if we don't use
C++17 as our projects standard. Hence, we just had dpp. Issues with this
arise when using MacOS dylib libraries, so we need to change it back to
the dpp::dpp target. The issue with C++ versions still persists, thus we
are forced to increase the C++ standard version to 20.