Added VCPKG toolchain file for Windows build
The Windows CMake configure was failing to find packages due to the
VCPKG toolchain that specifies paths for CMake not being used.

Also changed to workflow for Windows to not do the builds to get the
initial VCPKG cache so that it will run faster for debugging. I will
re-enable it after this workflow run.